Fixed a problem with the loop order of bombs
The order of bombs in the movement loop affected the outcome
of the collision handling. This was fixed by letting each bomb
only consider and fix collisions done with bombs of a higher
index than themself. This makes sure that each unique pair gets
one Has/Resolve check. Also the higher-index bomb will only get either
ignored or stopped, which means that there  is no need to check if
the bomb was moved to its target location this turn.

Very elegant, I have to say.. :3