Bomb collisions with blocking object is working properly
If a bomb moves right past another bomb (orthogonally), then nothing
happens.
However if the faster bomb is stopped by a static object, then the bombs
collide depending on the order they are evaluated.

Now all bombs are checked beforehand if they have static or agent
collisions
and if they do will reset their movement to IDLE. This leads to a target
position collision, which gets picked up by the subsequent loop

Note: There is an optimzation made here: Bombs can confidently say
they're
idle if their target position has an agent because there are only 2
critical
scenarios:

1) The agent has a destination position collision and bounces back.
However
since all agent movement is already done, this is not an issue.
2) The agent kicked a bomb that collides and bounces back. That is also
not a problem, because even if the agent bounces back there is still a
bomb at that position that prevents moving forward.